# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kadane Backend API
  description: Kadane Backend API 
  version: 1.0.0
servers:
  - url: https://api.kadane.xyz/v1
    description: Main production server
  - url: http://localhost:3000/v1
    description: Local development server
security: 
  -  Authorization: []
paths:
...
  /admin/problems/run:
    post:
      tags:
        - Admin
      summary: Run a problem
      description: Run a problem.
      operationId: runProblem
      requestBody:
        description: Problem data to be run
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProblemRunRequest'
      responses:
        '200':
          description: Problem run response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProblemRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /runs:
    post:
      tags:
        - Runs
      summary: Create a new run
      description: Create a new run for a specific problem.
      operationId: createRun
      requestBody:
        description: Run data to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
      responses:
        '201':
          description: Run created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /problems/{id}:
    get:
      tags:
        - Problems
      summary: Get problem by ID
      description: Retrieve a specific problem by its UUID
      operationId: getProblem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Problem UUID
      responses:
        '200':
          description: Problem retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Problem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes: 
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: jwt 
  schemas:
...
    StatusCode:
      type: string
      enum: [200, 201, 202, 204, 400, 401, 403, 404, 422, 500]
    
    # Responses
    # APIResponse is a generic response that can be used for any response
    # { "data": { any } }
...
    APIErrorData:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          anyOf:
            - $ref: '#/components/schemas/StatusCode'
    APIError:
      allOf:
        - $ref: '#/components/schemas/APIErrorData'
        - type: object
          properties:
            error:
              type: object
              $ref: '#/components/schemas/APIErrorData'
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/APIError'
        - type: object
          properties:
            code:
              type: string
              enum: [500]

    # Accounts
...
    AdminProblemRunRequest:
      description: Admin problem run request
      type: object
      properties:
        functionName:
          type: string
        solution:
          type: object
          additionalProperties:
            type: string
          description: Map of language identifiers to solution code
        testCase:
          $ref: '#/components/schemas/TestCase'
    TestCase:
      description: Test case
      type: object
      properties:
        description:
          type: string
        input:
          type: string
        output:
          type: string
        visibility:
          type: string
          enum: [public, private]
    AdminProblemRunResponse:
      description: Admin problem run response
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AdminProblemRunData'
    RunTestCase:
      description: Run test case
      type: object
      properties:
        time:
          type: string
        memory:
          type: integer
        status:
          type: string
        input:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseInput'
            nullable: true
        output:
          type: string
        compileOutput:
          type: string
        expectedOutput:
          type: string
...
    AdminProblemRunData:
      description: Admin problem run data
      type: object
      properties:
        runs:
          type: object
          properties:
            testCase:
              $ref: '#/components/schemas/RunTestCase'
            status:
              type: string
              enum: [accepted, wrong_answer, runtime_error, compilation_error, time_limit_exceeded, memory_limit_exceeded, output_limit_exceeded, internal_error]
            createdAt:
              type: string
              format: date-time
        status:
          type: string
          enum: [accepted, wrong_answer, runtime_error, compilation_error, time_limit_exceeded, memory_limit_exceeded, output_limit_exceeded, internal_error]
        completedAt:
          type: string
          format: date-time
...
    RunRequest:
      type: object
      properties:
        language:
          type: string
        sourceCode:
          type: string
        problemId:
          type: integer
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/TestCase'
    RunResult:
      type: object
      properties:
        id:
          type: string
        language:
          type: string
        time:
          type: string
          format: date-time
        memory:
          type: integer
          format: int32
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/RunTestCase'
        status:
          type: enum
          $ref: '#/components/schemas/SubmissionStatus'
        accountId:
          type: string
        problemId:
          type: integer
        createdAt:
          type: string
          format: date-time
    RunResultResponse:
      type: object
      properties:
        data:
          type: object
          $ref: '#/components/schemas/RunResult'
    # Solutions
...
    SubmissionStatus:
      type: string
      enum: [accepted, wrong_answer, runtime_error, compilation_error, time_limit_exceeded, memory_limit_exceeded, output_limit_exceeded, internal_error]
...
    Problem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        functionName:
          type: string
        tags:
          type: array
          items:
            type: string
        code:
          type: array
          items:
            $ref: '#/components/schemas/ProblemCode'
        hints:
          type: array
          items:
            $ref: '#/components/schemas/ProblemHint'
        points:
          type: integer
        testCases:
          type: array
          items:
            $ref: '#/components/schemas/ProblemTestCase'
        starred:
          type: boolean
        solved:
          type: boolean
        totalAttempts:
          type: integer
        totalCorrect:
          type: integer
    ProblemCode:
      type: object
      properties:
        language:
          type: string
        code:
          type: string
          format: byte
    ProblemHint:
      type: object
      properties:
        description:
          type: string
          format: byte
        answer:
          type: string
          format: byte
    ProblemTestCase:
      type: object
      properties:
        description:
          type: string
        input:
          type: array
          items:
            $ref: '#/components/schemas/TestCaseInput'
        output:
          type: string
        visibility:
          type: string
          enum: [public, private]
...
    TestCaseInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [int, float, double, string, boolean, array of int, array of float, array of double, array of string, array of boolean]
        value:
          type: string
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'